Struct wafel_viz::Vec4

#[repr(C)]
pub struct Vec4 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub w: f32,
}

A set of four coordinates which may be interpreted as a point or vector in 4d space, or as a homogeneous 3d vector or point.

Generally this distinction between a point and vector is more of a pain than it is worth to distinguish on a type level, however when converting to and from homogeneous coordinates it is quite important.

Fields
x: f32y: f32z: f32w: f32
Implementations

impl Vec4

pub const fn new(x: f32, y: f32, z: f32, w: f32) -> Vec4

pub const fn broadcast(val: f32) -> Vec4

pub fn unit_x() -> Vec4

pub fn unit_y() -> Vec4

pub fn unit_z() -> Vec4

pub fn unit_w() -> Vec4

pub fn dot(&self, other: Vec4) -> f32

pub fn reflect(&mut self, normal: Vec4)

pub fn reflected(&self, normal: Vec4) -> Vec4

pub fn mag_sq(&self) -> f32

pub fn mag(&self) -> f32

pub fn normalize(&mut self)

pub fn normalized(&self) -> Vec4

pub fn normalize_homogeneous_point(&mut self)

Normalize self in-place by interpreting it as a homogeneous point, i.e. scaling the vector to ensure the homogeneous component has length 1.

pub fn normalized_homogeneous_point(&self) -> Vec4

Normalize self by interpreting it as a homogeneous point, i.e. scaling the vector to ensure the homogeneous component has length 1.

pub fn truncated(&self) -> Vec3

Convert self into a Vec3 by simply removing its w component.

pub fn mul_add(&self, mul: Vec4, add: Vec4) -> Vec4

pub fn abs(&self) -> Vec4

pub fn clamp(&mut self, min: Vec4, max: Vec4)

pub fn clamped(self, min: Vec4, max: Vec4) -> Vec4

pub fn map<F>(&self, f: F) -> Vec4
where F: FnMut(f32) -> f32,

pub fn apply<F>(&mut self, f: F)
where F: FnMut(f32) -> f32,

pub fn max_by_component(self, other: Vec4) -> Vec4

pub fn min_by_component(self, other: Vec4) -> Vec4

pub fn component_max(&self) -> f32

pub fn component_min(&self) -> f32

pub fn zero() -> Vec4

pub fn one() -> Vec4

pub const fn xy(&self) -> Vec2

pub const fn xyz(&self) -> Vec3

pub fn layout() -> Layout

Get the core::alloc::Layout of Self

pub fn as_array(&self) -> &[f32; 4]

Interpret self as a statically-sized array of its base numeric type

pub fn as_mut_array(&mut self) -> &mut [f32; 4]

Interpret self as a statically-sized array of its base numeric type

pub fn as_slice(&self) -> &[f32]

Interpret self as a slice of its base numeric type

pub fn as_mut_slice(&mut self) -> &mut [f32]

Interpret self as a slice of its base numeric type

pub fn as_byte_slice(&self) -> &[u8]

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

pub const fn as_ptr(&self) -> *const f32

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

impl Vec4

pub fn refract(&mut self, normal: Vec4, eta: f32)

pub fn refracted(&self, normal: Vec4, eta: f32) -> Vec4