Crate wafel_viz

Structs

Camera

Defines the projection and view matrices for a scene.

LineElement

Mat4

A 4x4 square matrix.

PointElement

Rect2

Rect3

TriangleElement

Vec2

A set of two coordinates which may be interpreted as a vector or point in 2d space.

Vec3

A set of three coordinates which may be interpreted as a point or vector in 3d space, or as a homogeneous 2d vector or point.

Vec4

A set of four coordinates which may be interpreted as a point or vector in 4d space, or as a homogeneous 3d vector or point.

VizScene

Enums

Element

TransparencyHint

A hint for how to render a triangle with transparency, specifically to give wall hitboxes special behavior.

Viewport