wafel_viz
Struct Mat4

#[repr(C)]
pub struct Mat4 {
    pub cols: [Vec4; 4],
}

A 4x4 square matrix.

Useful for performing linear transformations (rotation, scaling) on 4d vectors, or for performing arbitrary transformations (linear + translation, projection, etc) on homogeneous 3d vectors.

Note that most constructors assume that the matrix will be used as a homogeneous 3d transformation matrix.

Fields
cols: [Vec4; 4]
Implementations

impl Mat4

pub const fn new(col1: Vec4, col2: Vec4, col3: Vec4, col4: Vec4) -> Mat4

pub fn identity() -> Mat4

pub fn from_translation(trans: Vec3) -> Mat4

Assumes homogeneous 3d coordinates.

pub fn from_scale(scale: f32) -> Mat4

Assumes homogeneous 3d coordinates.

pub fn from_nonuniform_scale(scale: Vec3) -> Mat4

Assumes homogeneous 3d coordinates.

pub fn from_scale_4d(scale: f32) -> Mat4

Full 4d diagonal matrix.

pub fn from_nonuniform_scale_4d(scale: Vec4) -> Mat4

Full 4d nonuniform scaling matrix.

pub fn from_euler_angles(roll: f32, pitch: f32, yaw: f32) -> Mat4

Angles are applied in the order roll -> pitch -> yaw

Roll is rotation inside the xy plane (“around the z axis”)

Pitch is rotation inside the yz plane (“around the x axis”)

Yaw is rotation inside the xz plane (“around the y axis”)

Assumes homogeneous 3d coordinates.

Important: This function assumes a right-handed, y-up coordinate space where:

+X axis points right

+Y axis points up

+Z axis points towards the viewer (i.e. out of the screen)

This means that you may see unexpected behavior when used with OpenGL or DirectX as they use a different coordinate system. You should use the appropriate projection matrix in projection module to fit your use case to remedy this.

pub fn from_rotation_x(angle: f32) -> Mat4

Create a new rotation matrix from a rotation “around the x axis”. This is here as a convenience function for users coming from other libraries; it is more proper to think of this as a rotation in the yz plane.

Assumes homogeneous 3d coordinates.

pub fn from_rotation_y(angle: f32) -> Mat4

Create a new rotation matrix from a rotation “around the y axis”. This is here as a convenience function for users coming from other libraries; it is more proper to think of this as a rotation in the xz plane.

Assumes homogeneous 3d coordinates.

pub fn from_rotation_z(angle: f32) -> Mat4

Create a new rotation matrix from a rotation “around the z axis”. This is here as a convenience function for users coming from other libraries; it is more proper to think of this as a rotation in the xy plane.

Assumes homogeneous 3d coordinates.

pub fn from_rotation_around(axis: Vec4, angle: f32) -> Mat4

Create a new rotation matrix from a rotation around the given axis. The axis will be interpreted as a 3d vector. This is here as a convenience function for users coming from other libraries.

Assumes homogeneous 3d coordinates.

pub fn from_angle_plane(angle: f32, plane: Bivec3) -> Mat4

Construct a rotation matrix given a bivector which defines a plane and rotation orientation, and a rotation angle.

plane must be normalized!

This is the equivalent of an axis-angle rotation.

Assumes homogeneous 3d coordinates.

pub fn translate(&mut self, translation: &Vec3)

Assumes homogeneous 3d coordinates.

pub fn translated(&self, translation: &Vec3) -> Mat4

Assumes homogeneous 3d coordinates.

pub fn look_at(eye: Vec3, at: Vec3, up: Vec3) -> Mat4

Constructs a ‘look-at’ matrix from an eye position, a focus position to look towards, and a vector that defines the ‘up’ direction.

This function assumes a right-handed, y-up coordinate space.

pub fn look_at_lh(eye: Vec3, at: Vec3, up: Vec3) -> Mat4

Constructs a ‘look-at’ matrix from an eye position, a focus position to look towards, and a vector that defines the ‘up’ direction.

This function assumes a left-handed, y-up coordinate space.

pub fn transpose(&mut self)

pub fn transposed(&self) -> Mat4

pub fn inverse(&mut self)

If this matrix is not currently invertable, this function will return an invalid inverse. This status is not checked by the library.

pub fn determinant(&self) -> f32

pub fn adjugate(&self) -> Mat4

The adjugate of this matrix, i.e. the transpose of the cofactor matrix.

This is equivalent to the inverse but without dividing by the determinant of the matrix, which can be useful in some contexts for better performance.

One such case is when this matrix is interpreted as a a homogeneous transformation matrix, in which case uniform scaling will not affect the resulting projected 3d version of transformed points or vectors.

pub fn inversed(&self) -> Mat4

If this matrix is not currently invertable, this function will return an invalid inverse. This status is not checked by the library.

pub fn transform_vec3(&self, vec: Vec3) -> Vec3

Transform a Vec3 by self, interpreting it as a vector.

pub fn transform_point3(&self, point: Vec3) -> Vec3

Transform a Vec3 by self, interpreting it as a point.

pub fn extract_translation(&self) -> Vec3

If self represents an affine transformation, return its translation components. Otherwise, the returned value has undefined properties.

pub fn extract_rotation(&self) -> Rotor3

If the 3x3 left upper block of self is a rotation, return the corresponding rotor. Otherwise, the returned value is a Rotor3 with undefined properties.

pub fn into_isometry(&self) -> Isometry3

If self represents an Isometry3 (i.e. self is a product of the from T * R where T is a translation and R a rotation), return the isometry

If self does not represent an isometry, the returned value has undefined properties.

pub fn truncate(&self) -> Mat3

Truncate self to a matrix consisting of the 3x3 left upper block. If you need a rotation, consider using Self::extract_rotation() instead.

pub fn layout() -> Layout

Get the core::alloc::Layout of Self

pub fn as_array(&self) -> &[f32; 16]

Interpret self as a statically sized array of the base numeric type.

pub fn as_mut_array(&mut self) -> &mut [f32; 16]

Interpret self as a statically sized array of the base numeric type.

pub fn as_component_array(&self) -> &[Vec4; 4]

Interpret self as a statically sized array of its component (column) vectors.

pub fn as_mut_component_array(&mut self) -> &mut [Vec4; 4]

Interpret self as a statically sized array of its component (column) vectors.

pub fn as_slice(&self) -> &[f32]

Interpret self as a slice of the base numeric type.

pub fn as_mut_slice(&mut self) -> &mut [f32]

Interpret self as a slice of the base numeric type.

pub fn as_component_slice(&self) -> &[Vec4]

Interpret self as a slice of the component (column) vectors

pub fn as_mut_component_slice(&mut self) -> &mut [Vec4]

Interpret self as a slice of the component (column) vectors

pub fn as_byte_slice(&self) -> &[u8]

Interpret self as a slice of bytes

pub fn as_mut_byte_slice(&mut self) -> &mut [u8]

Interpret self as a slice of bytes

pub const fn as_ptr(&self) -> *const f32

Returns a constant unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.

pub fn as_mut_ptr(&mut self) -> *mut f32

Returns a mutable unsafe pointer to the underlying data in the underlying type. This function is safe because all types here are repr(C) and can be represented as their underlying type.

Safety

It is up to the caller to correctly use this pointer and its bounds.